+++ src/Exception/ParseTemplateInteropException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace HeimrichHannot\NewsListReaderBundle\Exception;
+
+use Contao\Template;
+
+class ParseTemplateInteropException extends \Exception
+{
+    /** @noinspection PhpMissingParentConstructorInspection */
+    public function __construct(public readonly Template $template)
+    {
+    }
+}