+++ src/Resources/contao/dca/tl_reader_config.php
@@ -1,0 +1,5 @@
+<?php
+
+use HeimrichHannot\NewsListReaderBundle\DataContainer\DcaDefinitions;
+
+DcaDefinitions::addDefinitions('tl_reader_config');