+++ src/Resources/contao/languages/de/tl_reader_config.php
@@ -1,0 +1,6 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_reader_config'];
+
+$lang['useNewsExtension'] = ['Nachrichten-Erweiterung verwenden', 'Wählen Sie diese Option, wenn Sie die Nachrichten-Erweiterung verwenden möchten.'];
+$lang['news_imgSize'] = ['Nachrichten-Bildgröße', 'Hier können Sie die Abmessungen des Nachrichten-Bildes und den Skalierungsmodus festlegen.'];