+++ src/Resources/contao/languages/en/tl_list_config.php
@@ -1,0 +1,6 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config'];
+
+$lang['useNewsExtension'] = ['Use news extension', 'Use the news extension for this list.'];
+$lang['news_imgSize'] = ['News image size', 'Here you can set the dimensions of the news image and the scaling mode.'];